requirements.txt
python-dotenv
python-telegram-bot==13.15
kucoin
pybit
gate-api
mexc-sdk
